Select worlds when viewing the achievements screen too.
